Extract AvatarResolver
